+++ tests/apps/courses/test_factories_person.py
@@ -27,7 +27,7 @@ class PersonFactoryTestCase(TestCase):
         # The portrait plugins should point to one of our fixtures images
         for language in ["fr", "en"]:
             portrait_plugin = portrait.cmsplugin_set.get(
-                plugin_type="PicturePlugin", language=language
+                plugin_type="SimplePicturePlugin", language=language
             )
             self.assertIn(
                 "portrait",