+++ tests/apps/search/test_indexers_organizations.py
@@ -6,7 +6,6 @@ from unittest import mock
 from django.test import TestCase
 
 from cms.api import add_plugin
-from djangocms_picture.models import Picture
 
 from richie.apps.courses.factories import OrganizationFactory
 from richie.apps.search.indexers.organizations import OrganizationsIndexer
@@ -18,8 +17,9 @@ class OrganizationsIndexersTestCase(TestCase):
     and especially dynamic mapping shape in ES
     """
 
-    @mock.patch.object(
-        Picture, "img_src", new_callable=mock.PropertyMock, return_value="123.jpg"
+    @mock.patch(
+        "richie.apps.search.indexers.organizations.get_picture_info",
+        return_value="logo info",
     )
     def test_indexers_organizations_get_es_documents(self, _mock_picture):
         """
@@ -112,7 +112,7 @@ class OrganizationsIndexersTestCase(TestCase):
                         "en": "english description line 1. english description line 2.",
                         "fr": "description français ligne 1. description français ligne 2.",
                     },
-                    "logo": {"en": "123.jpg", "fr": "123.jpg"},
+                    "logo": {"en": "logo info", "fr": "logo info"},
                     "title": {
                         "en": "my first organization",
                         "fr": "ma première organisation",